Start phase 1 with initial bwgame decomposition
